src/requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-cors-headers==3.0.2
django-debug-toolbar==2.0
django-extensions==2.1.9
django-filter==2.1.0
django-filters==0.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-gis==0.14
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.3
pyrsistent==0.15.3
pytz==2019.1
requests==2.22.0
sentry-sdk==0.9.5
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
